+++ src/player.cpp
@@ -37,7 +37,7 @@ void Player::Reset()
 
 void Player::Draw()
 {
-    DrawTexture(texture, position.x, position.y, WHITE);
+    DrawTexture(texture, static_cast<int>(position.x), static_cast<int>(position.y), WHITE);
 }
 
 void Player::Update(float deltaT)
